Switch to targeting Java 7
As silly as this may sound, I've still got to compile against Java 7 for Jenkins plugins, at least for a while longer. I'd *really* like to use this library in a key Jenkins plugin, but can't unless it's compiled against Java 7. I tried this locally and everything worked fine. Thanks!

(Oh, and I also bumped the version to 0.1.5-SNAPSHOT)