Optimization for OneOf
An often-used pattern in conjunction with oneOf is to use a single value
enum as a constant where each oneOf alternative typically has a different
constant.

If the corresponding property is set in the json node (and thus is
textual) we can skip all alternatives that have a different constant
value defined for that property since they will definitely not match.

For schemas with many alternatives and a lot of items that need to run
through those alternatives, I found this to speedup validation
considerably (30% faster in my case).