Prevents recursive parsing of json schema when conditions like told in json-schema specification, point 8.3.1: https://json-schema.org/latest/json-schema-core.html#rfc.section.8.3.1
"A schema MUST NOT be run into an infinite loop against a schema. For example, if two schemas "#alice" and "#bob" both have an "allOf" property that refers to the other, a naive validator might get stuck in an infinite recursive loop trying to validate the instance. Schemas SHOULD NOT make use of infinite recursive nesting like this; the behavior is undefined. "