Split the PatternValidator into 2 classes.
This is necessary to guarantee that classes from org.joni and org.jcodings
are only loaded when the ecma262 configuration is set. Having both in the
same class and one of the class' members being a class from the before
mentioned packages fails with a NoClassDefFoundException when running in
an OSGI environments with the imports of the packages being optional and
not present.