Always normalize uri keys of JsonSchemaFactory.jsonMetaSchemas on both read and write.
Previously, when writing to the JsonSchemaFactory.jsonMetaSchemas hash map, the
uri key was not normalize, but upon reading from the hash map, the key was normalized.

In addition, if a key is not found in this hash map, the standard meta schema for the uri
is loaded.

The result was that an attempt to augment a standard meta schema with custom formatters did
not work because the standard meta schema was always used instead of the custom augmented
meta schema.

This change normalizes the uri keys of JsonSchemaFactory.jsonMetaSchemas on both read and write.
Normalizing the uri keys on both reads and writes also eliminates the need for the
forceHttps and removeEmptyFragmentSuffix arguments to the normalizeMetaSchemaUri() method.